+++ src/operators.jl
@@ -167,6 +167,9 @@ end
 (-)(q1::QuadExpr, q2::QuadExpr) = QuadExpr( vcat(q1.qvars1, q2.qvars1),     vcat(q1.qvars2, q2.qvars2),
                                             vcat(q1.qcoeffs, -q2.qcoeffs),  q1.aff - q2.aff)
 
+_deprecate_comparisons(sgn) =
+    Base.warn_once("The comparison operator $sgn has been deprecated for constructing constraints. Use the macro form @addConstraint instead.")
+
 # LinearConstraint
 # Number--???
 for (sgn, osgn) in ( (:<=,:>=), (:(==),:(==)), (:>=,:<=) )
@@ -179,9 +182,18 @@ for (sgn, osgn) in ( (:<=,:>=), (:(==),:(==)), (:>=,:<=) )
     end
 end
 # AffExpr--???
-(<=)(lhs::AffExpr, rhs::Number) = LinearConstraint(lhs,            -Inf,rhs-lhs.constant)
-(==)(lhs::AffExpr, rhs::Number) = LinearConstraint(lhs,rhs-lhs.constant,rhs-lhs.constant)
-(>=)(lhs::AffExpr, rhs::Number) = LinearConstraint(lhs,rhs-lhs.constant,             Inf)
+function (<=)(lhs::AffExpr, rhs::Number)
+    _deprecate_comparisons(:(<=))
+    LinearConstraint(lhs,-Inf,rhs-lhs.constant)
+end
+function (==)(lhs::AffExpr, rhs::Number)
+    _deprecate_comparisons(:(==))
+    LinearConstraint(lhs,rhs-lhs.constant,rhs-lhs.constant)
+end
+function (>=)(lhs::AffExpr, rhs::Number)
+    _deprecate_comparisons(:(>=))
+    LinearConstraint(lhs,rhs-lhs.constant,Inf)
+end
 for sgn in (:<=, :(==), :>=)
     for typ in (:Variable, :AffExpr, :QuadExpr)
         @eval $(sgn)(lhs::AffExpr, rhs::$(typ)) = $(sgn)(lhs-rhs, 0.0)
@@ -193,8 +205,12 @@ end
 # QuadConstraint
 # QuadConstraint--Number
 for sgn in (:<=, :(==), :>=)
-    @eval $(sgn)(lhs::QuadExpr, rhs::Number) =
-        QuadConstraint( QuadExpr(copy(lhs.qvars1), copy(lhs.qvars2), lhs.qcoeffs,lhs.aff - rhs), $(quot(sgn)))
+    @eval begin
+        function $(sgn)(lhs::QuadExpr, rhs::Number)
+            _deprecate_comparisons($sgn)
+            QuadConstraint( QuadExpr(copy(lhs.qvars1), copy(lhs.qvars2), lhs.qcoeffs,lhs.aff - rhs), $(quot(sgn)))
+        end
+    end
     for typ in (:Variable, :AffExpr, :QuadExpr)
         @eval $(sgn)(lhs::QuadExpr, rhs::$(typ)) = $(sgn)(lhs-rhs, 0)
     end
@@ -244,14 +260,14 @@ end
 #############################################################################
 # JuMPDict comparison operators (all errors)
 
-for sgn in (:<=, :(==), :>=)
-    for term in (:Real, :Variable, :AffExpr)
-        @eval begin
-            $(sgn)(a::JuMPContainer, b::$(term)) = error("Cannot construct constraint with a JuMPDict term")
-            $(sgn)(a::$(term), b::JuMPContainer) = error("Cannot construct constraint with a JuMPDict term")
-        end
-    end
-end
+# for sgn in (:<=, :(==), :>=)
+#     for term in (:Real, :Variable, :AffExpr)
+#         @eval begin
+#             $(sgn)(a::JuMPContainer, b::$(term)) = error("Cannot construct constraint with a JuMPDict term")
+#             $(sgn)(a::$(term), b::JuMPContainer) = error("Cannot construct constraint with a JuMPDict term")
+#         end
+#     end
+# end
 
 ###############################################################################
 # Add nonlinear function fallbacks for JuMP built-in types