+++ NEWS.md
@@ -1,6 +1,12 @@
 JuMP release notes
 ==================
 
+Unversioned
+-----------
+
+  * Comparison operators for constructing constraints (e.g. ``2x >= 1``) have been deprecated. Instead, construct the constraints explicitly in
+    the ``@addConstraint`` macro to add them to the model, or in the ``@LinearConstraint`` macro to create a stand-alone linear constraint instance.
+
 Version 0.8.0 (February 17, 2015)
 ---------------------------------